Source refactoring
- pages: Including pages of app
- services: Separate service
- components: Including common component
- config: routes,...
