+++ src/service/auth/types.ts
@@ -1,0 +1,4 @@
+
+export abstract class IAuth {
+    abstract signIn(username: string, password: string) : Promise<string>
+}
