+++ src/service/index.ts
@@ -1,4 +1,6 @@
 import {IAPI} from './types';
+import AuthService from "./auth";
+import TodoService from "./todo";
 
 let Service : IAPI;
 if (process.env.REACT_APP_WHOAMI === 'frontend') {
@@ -7,4 +9,5 @@ if (process.env.REACT_APP_WHOAMI === 'frontend') {
     Service = require('./api-fullstack').default as IAPI
 }
 
+export {AuthService,TodoService};
 export default Service
