+++ src/service/todo/index.tsx
@@ -1,0 +1,22 @@
+import {Todo, TodoStatus} from "models/todo";
+import shortid from "shortid";
+import {ITodo} from "./types";
+
+
+class TodoService extends ITodo {
+    async createTodo(content: string): Promise<Todo> {
+        return Promise.resolve({
+            content: content,
+            created_date: new Date().toISOString(),
+            status: TodoStatus.ACTIVE,
+            id: shortid(),
+            user_id: 'firstUser'
+        } as Todo);
+    }
+
+    async getTodos(): Promise<Todo[]> {
+        return []
+    }
+}
+
+export default new TodoService();
