+++ src/service/todo/types.ts
@@ -1,0 +1,6 @@
+import {Todo} from "models/todo";
+
+export abstract class ITodo {
+    abstract getTodos() : Promise<Array<Todo>>
+    abstract createTodo(content: string) : Promise<Todo>
+}
