+++ src/service/types.ts
@@ -1,4 +1,4 @@
-import {Todo} from '../models/todo';
+import {Todo} from 'models/todo';
 
 export abstract class IAPI {
     abstract signIn(username: string, password: string) : Promise<string>
