+++ src/store/actions/todos.ts
@@ -1,4 +1,4 @@
-import {Todo} from "../models/todo";
+import {Todo} from "models/todo";
 
 export const SET_TODO = 'SET_TODO';
 export const CREATE_TODO = 'CREATE_TODO';
